move admission integration tests to separate file
